Restructure all files
